requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

# MongoDB + ODM
beanie==1.26.0
motor==3.5.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment variables (optional but useful)
python-dotenv==1.0.1

# Validation & typing (FastAPI uses these)
pydantic==2.8.2
pydantic-settings==2.5.2

python-dotenv==1.0.1

authlib==1.3.0
httpx==0.27.0
pydantic==2.7.1